Microservices_django/Post_comment_with_docker/Post/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.7
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2
